CRUD_Modifie

+++ flask_base/src/services/songs.py
@@ -1,6 +1,7 @@
 # flask/services/songs.py
 import json
 import requests
+import uuid
 from sqlalchemy import exc
 from marshmallow import EXCLUDE
 from flask_login import current_user
@@ -17,43 +18,76 @@ from src.schemas.errors import *
 songs_url = "http://localhost:8092/songs/"
 
 
+
+def isSongIdValid(id):
+    verif = requests.request(method="GET", url=songs_url+(id))
+    if verif.status_code == 200:
+        return True
+    else:
+        return False
+
+
+
+def isUuidValid(id):
+    try:
+        uuid_obj = uuid.UUID(id)
+        return uuid_obj.version == 4
+    except ValueError:
+        # La conversion en UUID a échoué
+        return False
+
+
+
+def isUserIdValid(id):
+    verif = requests.request(method="GET", url=songs_url+id)
+    if verif.json()["idUser"] == current_user.id:
+        return True
+    else:
+        return False
+
+
 def get_song(id):
+
+    if not isUuidValid(id):
+        raise UnprocessableEntity
+    if not isSongIdValid(id):
+        raise NotFound
+
     response = requests.get(songs_url + id)
+
+    if response.status_code == 500:
+        raise SomethingWentWrong  
     return response.json(), response.status_code
 
 
 def create_song(song_data):
+    
     song_schema = CreateSongSchema().loads(json.dumps(song_data), unknown=EXCLUDE)
     response = requests.request(method="POST", url=songs_url, json=song_schema)
-
-    if response.status_code != 201:
-        return response.json(), response.status_code
-
-    # Ajouter la chanson dans la base de données
-    try:
-        song_model = SongSchema().load(response.json())
-        songs_repository.create_song(song_model)
-    except Exception:
+    
+    if response.status_code == 500:
         raise SomethingWentWrong
 
-    return response.json(), response.status_code
+    return "", response.status_code
 
 
+def get_song_from_db(song_id):
+    return songs_repository.get_song(song_id)
 
-def update_song(id, updated_data):
-    song_schema = UpdateSongSchema().loads(json.dumps(updated_data), unknown=EXCLUDE)
-    response = requests.request(method="PUT", url=songs_url + id, json=song_schema)
 
-    if response.status_code != 200:
-        return response.json(), response.status_code
+def song_exists(song_id):
+    return get_song_from_db(song_id) is not None
 
-    # Mettre à jour la chanson dans la base de données
-    try:
-        updated_song_model = SongSchema().load(response.json())
-        songs_repository.update_song(updated_song_model)
-    except Exception:
+def update_song(id, song_update):
+    if not isUuidValid(id):
+        raise UnprocessableEntity
+    if not isSongIdValid(id):
+        raise NotFound
+
+    response = requests.request(method="PUT", url=songs_url+id,json=song_update)
+
+    if response.status_code == 500:
         raise SomethingWentWrong
-
     return response.json(), response.status_code
 
 
@@ -64,12 +98,12 @@ def get_songs():
 
 
 def delete_song(id):
+    if not isSongIdValid(id):
+        raise NotFound
+   
     response=requests.request(method="DELETE", url=songs_url + id)
 
-    try:
-        songs_repository.delete_song(id)
-    except Exception:
-        error = NotFoundSchema().loads("{}")
-        return error, error.get("code")
+    if response.status_code == 500:
+        raise SomethingWentWrong
 
-    return response.json(), 200
+    return "", response.status_code
